requirements.txt
boto3
constructs>=10.0.0,<11.0.0
aws-cdk-lib==2.179.0